Separate events into it’s own library
