+++ audit/package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@libero/auth-utils": "../lib/auth-utils",
     "@libero/event-bus": "../lib/event-bus",
+    "@libero/events": "../lib/events",
     "dotenv-cli": "^2.0.1",
     "express": "^4.17.1",
     "funfix": "^7.0.1",
