+++ audit/src/handlers/index.test.ts
@@ -1,7 +1,7 @@
 import { Event } from '@libero/event-bus';
 import { ServiceStartedHandler, UserLoggedInHandler } from './index';
 import { AuditController } from '../domain/audit';
-import { ServiceStartedPayload, serviceStartedIdentifier, userLoggedInIdentifier, UserLoggedInPayload } from '../events';
+import { ServiceStartedPayload, serviceStartedIdentifier, userLoggedInIdentifier, UserLoggedInPayload } from '@libero/events';
 
 jest.mock('../logger');
