+++ audit/src/handlers/index.ts
@@ -1,6 +1,6 @@
 // Event handlers - returns
 import { Event } from '@libero/event-bus';
-import { ServiceStartedPayload, UserLoggedInPayload } from '../events';
+import { ServiceStartedPayload, UserLoggedInPayload } from '@libero/events';
 import { InfraLogger as logger } from '../logger';
 import { AuditController } from '../domain/audit';
 import { AuditLogItem } from '../domain/types';
