+++ lib/Dockerfile.packages
@@ -7,3 +7,4 @@ COPY . .
 # Yeah I know this isn't perfect but this build stage is pretty much throwaway - we just take the lib folder and use that
 RUN ["sh", "-c", "cd auth-utils && yarn && yarn build"]
 RUN ["sh", "-c", "cd event-bus && yarn && yarn build"]
+RUN ["sh", "-c", "cd events && yarn && yarn build"]
